# doc/api.yaml
openapi: 3.0.3
info:
  title: WASA-Exam
  description: API for the WASA Exam project.
  version: 1.0.0
security: []
...
paths:
...
  /conversations/{conversationId}/messages/{messageId}/reaction:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
      - name: messageId
        in: path
        required: true
        description: The ID of the message
        schema:
          type: string
          description: UUID of message
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    put:
      tags: [Reaction]
      summary: Add a reaction to a message
      description: Add a reaction (emoji) to a message.
      operationId: commentMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Reaction details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Reaction payload
              required: [type]
              properties:
                type:
                  type: string
                  description: Type of reaction to add
                  enum: [like, heart, laugh, sad_face, angry_face]
      responses:
        "200":
          description: Reaction added successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Reaction added"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Invalid reaction"
        "404":
          description: Message or conversation not found
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Message not found"

    delete:
      tags: [Reaction]
      summary: Remove a reaction from a message
      description: Remove a reaction from a message.
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          description: The type of reaction to remove
          schema:
            type: string
            description: Reaction type
            enum: [like, heart, laugh, sad_face, angry_face]
      responses:
        "200":
          description: Reaction removed successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    example: "Reaction removed"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                description: Error response
                properties:
                  error:
                    type: string
                    description: Error message
        "404":
          description: Message or conversation not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...